Set up main menu bar and infrastructure of the app
Added BaseActivity.java, an abstract class which inherits AppCompatActivity and sets up the behaviour of the main menu bar.
Added MainDrawerFragment.java which inflates navigation options as a fragment at the bottom of the screen onCreateView (when left button is pressed in the menu).
Added xml files to display the main menu and changed activity_main.xml to include activity_base.xml.
Changed MainActivity.java to inherit BaseActivity.java.